requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
pyodbc==4.0.39

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing & Validation
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2

# Report Generation
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Task Scheduling & Background Jobs
celery==5.3.4
redis==5.0.1

# HTTP Client & Web Utilities
httpx==0.25.2
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0

# Date/Time Handling
python-dateutil==2.8.2

# Image Processing (for signatures/reports)
pillow==10.1.0

# Optional: Additional libraries for enhanced functionality
# For Excel file processing beyond openpyxl
xlwings==0.31.2

# For advanced PDF features
PyPDF2==3.0.1

# For email notifications (if needed)
aiosmtplib==2.0.2

# For file handling
aiofiles==23.2.1

openpyxl